+++ TerminalCommander/Patches/GameManagement.cs
@@ -1,12 +1,7 @@
 ﻿using BepInEx.Logging;
 using HarmonyLib;
+using LethalNetworkAPI;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TerminalApi.Classes;
-using Unity.Netcode;
 
 namespace TerminalCommander.Patches
 {
@@ -15,12 +10,15 @@ namespace TerminalCommander.Patches
     {
         private static ManualLogSource logSource; // Log source field
         private static Commander commanderSource;
-
+        private static LethalClientMessage<TerminalCommanderConfiguration> clientMessage = new LethalClientMessage<TerminalCommanderConfiguration>("config");
+        private static LethalServerMessage<TerminalCommanderConfiguration> serverMessage = new LethalServerMessage<TerminalCommanderConfiguration>("config");
         // Method to set the log source
         public static void SetSource(Commander source)
         {
             commanderSource = source;
             logSource = commanderSource.log;
+            clientMessage.OnReceived += CustomClientMessage_OnReceived;
+            serverMessage.OnReceived += ServerMessage_OnReceived;
         }
 
         [HarmonyPatch("OnClientConnect")]
@@ -29,45 +27,30 @@ namespace TerminalCommander.Patches
         {
             try
             {
-                if (!__instance.IsServer)
+                if(__instance.IsHost)
                 {
-                    return;
+                    logSource.LogInfo($"{Commander.modName} syncing configurations for connected player: clientId {clientId}.");
+                    serverMessage.SendAllClients(commanderSource.Configs);
                 }
-
-                //Sync Configs
-                logSource.LogInfo($"{Commander.modName} syncing configurations for new player.");
-                commanderSource.NetworkHandler.SyncConfigs(commanderSource.Configs);
-
             }
             catch (Exception ex)
             {
                 //Configs not synced
-                logSource.LogInfo($"{Commander.modName} GAME MANAGEMENT ERROR (CONFIGS NOT SYNCED): {ex.Message}");
+                logSource.LogError($"GAME MANAGEMENT ERROR (CONFIGS NOT SYNCED): {ex.Message}");
             };
         }
-
-        [HarmonyPatch("PlayerLoadedServerRpc")]
-        [HarmonyPostfix]
-        static void PushHostConfigurationPatch(StartOfRound __instance, ulong clientId)
+        private static void CustomClientMessage_OnReceived(TerminalCommanderConfiguration obj)
         {
-            try
-            {
-                if (!__instance.IsServer)
-                {
-                    return;
-                }
-
-                //Sync Configs
-                logSource.LogInfo($"{Commander.modName} syncing configurations to server.");
-                commanderSource.NetworkHandler.SyncConfigs(commanderSource.Configs);
-
-            }
-            catch (Exception ex)
-            {
-                //Configs not synced
-                logSource.LogInfo($"{Commander.modName} GAME MANAGEMENT ERROR (CONFIGS NOT SYNCED TO SERVER): {ex.Message}");
-            };
+            logSource.LogInfo($"Host configurations received.");
+            commanderSource.Configs.AllowBigDoors = obj.AllowBigDoors;
+            commanderSource.Configs.AllowJamming = obj.AllowJamming;
         }
 
+        private static void ServerMessage_OnReceived(TerminalCommanderConfiguration arg1, ulong arg2)
+        {
+            logSource.LogInfo($"Sending host configurations to clients.");
+            clientMessage.SendServer(commanderSource.Configs);
+        }
     }
+
 }